feat: search header 컴포넌트에서 region 데이터 fetch 및 search input 컴포넌트 분리 (#160)
