+++ apps/web/components/layout/header/SearchHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchHeader } from './SearchHeader';
