+++ apps/web/components/layout/header/index.ts
@@ -1,2 +1,2 @@
 export { default as TopNavigation } from './TopNavigation';
-export { default as SearchHeader } from './SearchHeader';
+export { SearchHeader } from './SearchHeader';
